Add throw an error in firebase storage file deletion if url passed is null

+++ src/api/v1/infrastructure/services/FirebaseStorageServices.js
@@ -29,6 +29,9 @@ module.exports = class FirebaseStorageService {
   }
 
   async delete(url) {
+    if (!url) {
+      throw new Error(`URL can't be null`);
+    }
     const response = await this.storage.bucket().deleteFiles({prefix: url});
     return `Success deleted: ${url}`;
   }
